Update whois.cira.ca.rb
This parser is incorrectly interpreting a key with no accompanying value as a header.  An example of this can currently be seen in olivegarden.ca, where "Fax:" has no value and as a result the "Email:" value that follows it is no properly inserted into the associated contact.  The fix relies on the lack of indentation for a line to be considered a header.